backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# AI/LLM
transformers==4.35.2
torch==2.1.1
accelerate==0.25.0
bitsandbytes==0.41.3
sentencepiece==0.1.99

# Crawling
requests==2.31.0
beautifulsoup4==4.12.2
praw==7.7.1

# Task Scheduling
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-slugify==8.0.1
Werkzeug==3.0.1

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Optional
redis==5.0.1
celery==5.3.4
